Sort game autocomplete options alphabetically
