+++ src/commands/rateSingleGame.ts
@@ -11,6 +11,7 @@ import { Game } from "@prisma/client";
 
 import { Command, prisma } from "..";
 
+import { gameAutocomplete } from "../util/gameAutocomplete";
 import { registerUser } from "../util/registerUser";
 
 const gameEmbedBuilder = (game: Game): APIEmbed => {
@@ -152,17 +153,5 @@ export const rateSingleGame: Command = {
     }
   },
 
-  autocomplete: async (interaction) => {
-    const guildId = interaction.guildId;
-    if (!guildId) return;
-
-    const focusedValue = interaction.options.getFocused();
-    const games = await prisma.game.findMany({
-      where: { guildId, name: { contains: focusedValue, mode: "insensitive" } },
-    });
-
-    await interaction.respond(
-      games.map(game => ({ name: game.name, value: game.id })),
-    );
-  },
+  autocomplete: gameAutocomplete,
 };
